refactor: extract lenient json parser
