+++ src/main/java/berlin/yuna/typemap/logic/JsonDecoder.java
@@ -12,29 +12,19 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.PushbackReader;
 import java.io.Reader;
 import java.io.UncheckedIOException;
 import java.io.PushbackInputStream;
-import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URL;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Iterator;
-import java.util.Spliterator;
-import java.util.Spliterators;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Consumer;
 import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 
-import static berlin.yuna.typemap.logic.JsonEncoder.unescapeJson;
-import static berlin.yuna.typemap.logic.TypeConverter.convertObj;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Collections.singletonList;
-import static java.util.Spliterators.spliteratorUnknownSize;
 
 public class JsonDecoder {
 
@@ -186,57 +176,7 @@ public class JsonDecoder {
      * @return lazy stream of Type elements (caller must close)
      */
     public static Stream<Object> streamArray(final InputStream json, final Charset charset) throws IOException {
-        final LenientStream stream = new LenientStream(new InputStreamReader(json, charset), 8 * 1024);
-        final int start = stream.nextNonWhitespace();
-        if (start != '[') {
-            throw new IllegalStateException("Expected array start '['");
-        }
-
-        final Iterator<Object> iterator = new Iterator<>() {
-            private boolean endReached;
-
-            @Override
-            public boolean hasNext() {
-                if (endReached) {
-                    return false;
-                }
-                try {
-                    final int next = stream.nextNonWhitespace();
-                    if (next == ']') {
-                        endReached = true;
-                        return false;
-                    }
-                    stream.unread(next);
-                    return true;
-                } catch (final IOException e) {
-                    endReached = true;
-                    return false;
-                }
-            }
-
-            @Override
-            public Object next() {
-                if (!hasNext()) {
-                    throw new java.util.NoSuchElementException();
-                }
-                try {
-                    final Object value = parseValue(stream);
-                    final int sep = stream.nextNonWhitespace();
-                    if (sep == ']') {
-                        endReached = true;
-                    } else if (sep != ',') {
-                        throw new IllegalStateException("Invalid array separator");
-                    }
-                    return value;
-                } catch (final IOException e) {
-                    endReached = true;
-                    throw new IllegalStateException(e);
-                }
-            }
-        };
-
-        return StreamSupport.stream(spliteratorUnknownSize(iterator, Spliterator.ORDERED), false)
-                .onClose(() -> closeQuietly(stream.reader));
+        return JsonLenientParser.streamArray(new InputStreamReader(json, charset));
     }
 
     /**
@@ -248,64 +188,7 @@ public class JsonDecoder {
      */
     @SuppressWarnings("java:S3776")
     public static Stream<Pair<String, Object>> streamObject(final InputStream json, final Charset charset) throws IOException {
-        final LenientStream stream = new LenientStream(new InputStreamReader(json, charset), 8 * 1024);
-        final int start = stream.nextNonWhitespace();
-        if (start != '{')
-            throw new IllegalStateException("Expected object start '{'");
-        final Iterator<Pair<String, Object>> iterator = new Iterator<>() {
-            private boolean endReached;
-
-            @Override
-            public boolean hasNext() {
-                if (endReached) {
-                    return false;
-                }
-                try {
-                    final int next = stream.nextNonWhitespace();
-                    if (next == '}') {
-                        endReached = true;
-                        return false;
-                    }
-                    stream.unread(next);
-                    return true;
-                } catch (final IOException e) {
-                    endReached = true;
-                    return false;
-                }
-            }
-
-            @Override
-            public Pair<String, Object> next() {
-                if (!hasNext()) {
-                    throw new java.util.NoSuchElementException();
-                }
-                try {
-                    final int first = stream.nextNonWhitespace();
-                    if (first == '}') {
-                        endReached = true;
-                        return null;
-                    }
-                    final String key = parseKey(stream, first);
-                    final int sep = stream.nextNonWhitespace();
-                    if (sep != ':') {
-                        throw new IllegalStateException("Invalid JSON object separator");
-                    }
-                    final Object value = parseValue(stream);
-                    final int endOrComma = stream.nextNonWhitespace();
-                    if (endOrComma == '}') {
-                        endReached = true;
-                    } else if (endOrComma != ',') {
-                        throw new IllegalStateException("Invalid JSON object separator");
-                    }
-                    return new Pair<>(key, value);
-                } catch (final IOException e) {
-                    endReached = true;
-                    throw new IllegalStateException(e);
-                }
-            }
-        };
-
-        return errorTolerantStream(iterator, () -> closeQuietly(stream.reader));
+        return JsonLenientParser.streamObject(new InputStreamReader(json, charset));
     }
 
     /**
@@ -318,7 +201,7 @@ public class JsonDecoder {
     public static Stream<Pair<Integer, Object>> streamJsonArray(final InputStream json, final Charset charset) {
         try {
             final AtomicInteger index = new AtomicInteger(0);
-            return errorTolerantStream(streamArray(json, charset).map(value -> new Pair<>(index.getAndIncrement(), value)));
+            return JsonLenientParser.errorTolerantStream(streamArray(json, charset).map(value -> new Pair<>(index.getAndIncrement(), value)));
         } catch (final Exception ignored) {
             return Stream.empty();
         }
@@ -350,11 +233,7 @@ public class JsonDecoder {
     public static Stream<Pair<Integer, Object>> streamJsonArray(final Path path, final Charset charset) {
         if (path == null)
             return Stream.empty();
-        try {
-            return streamJsonArray(Files.newInputStream(path), charset);
-        } catch (final IOException e) {
-            throw new UncheckedIOException(e);
-        }
+        return open(() -> Files.newInputStream(path), charset, JsonDecoder::streamJsonArray);
     }
 
     public static Stream<Pair<Integer, Object>> streamJsonArray(final Path path) {
@@ -374,11 +253,7 @@ public class JsonDecoder {
     public static Stream<Pair<Integer, Object>> streamJsonArray(final URI uri, final Charset charset) {
         if (uri == null)
             return Stream.empty();
-        try {
-            return streamJsonArray(uri.toURL(), charset);
-        } catch (final MalformedURLException e) {
-            throw new UncheckedIOException(new IOException(e));
-        }
+        return open(() -> uri.toURL().openStream(), charset, JsonDecoder::streamJsonArray);
     }
 
     public static Stream<Pair<Integer, Object>> streamJsonArray(final URI uri) {
@@ -388,11 +263,7 @@ public class JsonDecoder {
     public static Stream<Pair<Integer, Object>> streamJsonArray(final URL url, final Charset charset) {
         if (url == null)
             return Stream.empty();
-        try {
-            return streamJsonArray(url.openStream(), charset);
-        } catch (final IOException e) {
-            throw new UncheckedIOException(e);
-        }
+        return open(url::openStream, charset, JsonDecoder::streamJsonArray);
     }
 
     public static Stream<Pair<Integer, Object>> streamJsonArray(final URL url) {
@@ -404,7 +275,7 @@ public class JsonDecoder {
      */
     public static Stream<Pair<String, Object>> streamJsonObject(final InputStream json, final Charset charset) {
         try {
-            return errorTolerantStream(streamObject(json, charset));
+            return streamObject(json, charset);
         } catch (final Exception ignored) {
             return Stream.empty();
         }
@@ -436,11 +307,7 @@ public class JsonDecoder {
     public static Stream<Pair<String, Object>> streamJsonObject(final Path path, final Charset charset) {
         if (path == null)
             return Stream.empty();
-        try {
-            return streamJsonObject(Files.newInputStream(path), charset);
-        } catch (final IOException e) {
-            throw new UncheckedIOException(e);
-        }
+        return open(() -> Files.newInputStream(path), charset, JsonDecoder::streamJsonObject);
     }
 
     public static Stream<Pair<String, Object>> streamJsonObject(final Path path) {
@@ -460,11 +327,7 @@ public class JsonDecoder {
     public static Stream<Pair<String, Object>> streamJsonObject(final URI uri, final Charset charset) {
         if (uri == null)
             return Stream.empty();
-        try {
-            return streamJsonObject(uri.toURL(), charset);
-        } catch (final MalformedURLException e) {
-            throw new UncheckedIOException(new IOException(e));
-        }
+        return open(() -> uri.toURL().openStream(), charset, JsonDecoder::streamJsonObject);
     }
 
     public static Stream<Pair<String, Object>> streamJsonObject(final URI uri) {
@@ -474,11 +337,7 @@ public class JsonDecoder {
     public static Stream<Pair<String, Object>> streamJsonObject(final URL url, final Charset charset) {
         if (url == null)
             return Stream.empty();
-        try {
-            return streamJsonObject(url.openStream(), charset);
-        } catch (final IOException e) {
-            throw new UncheckedIOException(e);
-        }
+        return open(url::openStream, charset, JsonDecoder::streamJsonObject);
     }
 
     public static Stream<Pair<String, Object>> streamJsonObject(final URL url) {
@@ -501,7 +360,7 @@ public class JsonDecoder {
             final PushbackInputStream pushback = json instanceof final PushbackInputStream p ? p : new PushbackInputStream(json, 1);
             final int ch = nextNonWhitespace(pushback);
             if (ch == -1) {
-                closeQuietly(pushback);
+                JsonLenientParser.closeQuietly(pushback);
                 return Stream.empty();
             }
             pushback.unread(ch);
@@ -510,13 +369,13 @@ public class JsonDecoder {
                     return (Stream<Pair<Object, Object>>) (Stream<?>) streamJsonObject(pushback, charset);
                 if (ch == '[')
                     return (Stream<Pair<Object, Object>>) (Stream<?>) streamJsonArray(pushback, charset);
-                closeQuietly(pushback);
+                JsonLenientParser.closeQuietly(pushback);
                 throw new UncheckedIOException(new IOException("Unsupported top-level token: " + (char) ch));
             } catch (final UncheckedIOException e) {
-                closeQuietly(pushback);
+                JsonLenientParser.closeQuietly(pushback);
                 throw e;
             } catch (final Exception e) {
-                closeQuietly(pushback);
+                JsonLenientParser.closeQuietly(pushback);
                 throw new UncheckedIOException(new IOException(e));
             }
         } catch (final IOException e) {
@@ -569,11 +428,7 @@ public class JsonDecoder {
         if (path == null) {
             return Stream.empty();
         }
-        try {
-            return streamJson(Files.newInputStream(path), charset);
-        } catch (final IOException e) {
-            throw new UncheckedIOException(e);
-        }
+        return open(() -> Files.newInputStream(path), charset, JsonDecoder::streamJson);
     }
 
     /**
@@ -607,11 +462,7 @@ public class JsonDecoder {
         if (uri == null) {
             return Stream.empty();
         }
-        try {
-            return streamJson(uri.toURL(), charset);
-        } catch (final MalformedURLException e) {
-            throw new UncheckedIOException(new IOException(e));
-        }
+        return open(() -> uri.toURL().openStream(), charset, JsonDecoder::streamJson);
     }
 
     /**
@@ -627,11 +478,7 @@ public class JsonDecoder {
     public static Stream<Pair<Object, Object>> streamJson(final URL url, final Charset charset) {
         if (url == null)
             return Stream.empty();
-        try {
-            return streamJson(url.openStream(), charset);
-        } catch (final IOException e) {
-            throw new UncheckedIOException(e);
-        }
+        return open(url::openStream, charset, JsonDecoder::streamJson);
     }
 
     /**
@@ -641,9 +488,40 @@ public class JsonDecoder {
         return streamJson(url, UTF_8);
     }
 
+    public interface InputSupplier {
+        InputStream open() throws IOException;
+    }
+
+    public interface StreamFactory<T> {
+        Stream<T> open(InputStream input, Charset charset) throws IOException;
+    }
+
+    public static <T> Stream<T> open(final InputSupplier supplier, final Charset charset, final StreamFactory<T> factory) {
+        if (supplier == null) {
+            return Stream.empty();
+        }
+        try {
+            return open(supplier.open(), charset, factory);
+        } catch (final IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
+
+    public static <T> Stream<T> open(final InputStream input, final Charset charset, final StreamFactory<T> factory) {
+        if (input == null) {
+            return Stream.empty();
+        }
+        try {
+            return factory.open(input, charset);
+        } catch (final IOException e) {
+            JsonLenientParser.closeQuietly(input);
+            throw new UncheckedIOException(e);
+        }
+    }
+
     /**
      * Detects whether the input stream is XML (leading '<') or JSON and delegates to the appropriate parser.
-     * Uses streaming to avoid buffering whole payloads; wraps incoming streams in a {@link BufferedInputStream} if needed.
+     * Uses streaming to avoid buffering whole payloads; wraps incoming streams in a {@link PushbackInputStream}.
      *
      * @param input   stream to parse
      * @param charset charset to use when decoding JSON content
@@ -664,60 +542,8 @@ public class JsonDecoder {
                 return XmlDecoder.xmlTypeOf(pushback);
             }
             try (final Reader reader = buffered(new InputStreamReader(pushback, charset))) {
-                return parse(reader);
+                return JsonLenientParser.parse(reader);
             }
-        }
-    }
-
-    private static void closeQuietly(final AutoCloseable closeable) {
-        if (closeable != null) {
-            try {
-                closeable.close();
-            } catch (final Exception ignored) {
-                // ignore
-            }
-        }
-    }
-
-    private static <T> Stream<T> errorTolerantStream(final Iterator<T> iterator, final Runnable onClose) {
-        final Spliterator<T> spliterator = new Spliterators.AbstractSpliterator<>(Long.MAX_VALUE, Spliterator.ORDERED) {
-            @Override
-            public boolean tryAdvance(final Consumer<? super T> action) {
-                try {
-                    if (iterator.hasNext()) {
-                        action.accept(iterator.next());
-                        return true;
-                    }
-                    return false;
-                } catch (final Exception e) {
-                    throw new UncheckedIOException(new IOException(e));
-                }
-            }
-        };
-        return StreamSupport.stream(spliterator, false).onClose(onClose);
-    }
-
-    private static <T> Stream<T> errorTolerantStream(final Stream<T> input) {
-        try {
-            return errorTolerantStream(input.iterator(), input::close);
-        } catch (final Exception ignored) {
-            closeQuietly(input);
-            return Stream.empty();
-        }
-    }
-
-    private static Object parse(final Reader reader) throws IOException {
-        final LenientStream stream = new LenientStream(reader, 8 * 1024);
-        try {
-            final Object value = parseValue(stream);
-            if (value instanceof final LinkedTypeMap singleMap && singleMap.size() == 1 && singleMap.containsKey(""))
-                return singleMap.get("");
-            if (value instanceof final LinkedTypeMap map && map.isEmpty() && "{}".contentEquals(stream.rawText().strip()))
-                return null;
-            return value;
-        } catch (final RuntimeException ignored) {
-            final String raw = stream.rawText();
-            return raw.isEmpty() ? null : raw;
         }
     }
 
@@ -731,221 +557,6 @@ public class JsonDecoder {
             ch = input.read();
         } while (ch != -1 && Character.isWhitespace(ch));
         return ch;
-    }
-
-    private static Object parseValue(final LenientStream stream) throws IOException {
-        final int ch = stream.nextNonWhitespace();
-        if (ch == -1) {
-            return null;
-        }
-        return switch (ch) {
-            case '{' -> parseObject(stream);
-            case '[' -> parseArray(stream);
-            case '"' -> parseString(stream);
-            default -> parsePrimitive(stream, (char) ch);
-        };
-    }
-
-    @SuppressWarnings("java:S3776")
-    private static LinkedTypeMap parseObject(final LenientStream stream) throws IOException {
-        final LinkedTypeMap map = new LinkedTypeMap();
-        while (true) {
-            final int first = stream.nextNonWhitespace();
-            if (first == '}') {
-                return map;
-            } else if (first == -1) {
-                throw new IllegalStateException("Unterminated object");
-            }
-            final String key = parseKey(stream, first);
-            final int separator = stream.nextNonWhitespace();
-            if (separator == ':') {
-                final Object value = parseValue(stream);
-                map.put(key, value);
-                final int sep = stream.nextNonWhitespace();
-                if (sep == '}') {
-                    return map;
-                } else if (sep != ',') {
-                    throw new IllegalStateException("Invalid JSON object separator");
-                }
-            } else if (separator == '}' || separator == ',') {
-                map.put("", key);
-                if (separator == '}') {
-                    return map;
-                }
-            } else {
-                throw new IllegalStateException("Invalid JSON object separator");
-            }
-        }
-    }
-
-    private static String parseKey(final LenientStream stream, final int first) throws IOException {
-        if (first == '"') {
-            return parseString(stream);
-        }
-        final StringBuilder token = new StringBuilder();
-        token.append((char) first);
-        while (true) {
-            final int ch = stream.read();
-            if (ch == -1) {
-                return token.toString();
-            }
-            if (ch == ':') {
-                stream.unread(ch);
-                return token.toString();
-            }
-            if (!Character.isWhitespace(ch)) {
-                token.append((char) ch);
-            }
-        }
-    }
-
-    private static TypeList parseArray(final LenientStream stream) throws IOException {
-        final TypeList list = new TypeList();
-        while (true) {
-            int ch = stream.nextNonWhitespace();
-            if (ch == ']') {
-                return list;
-            }
-            if (ch == -1) {
-                throw new IllegalStateException("Unterminated array");
-            }
-            stream.unread(ch);
-            list.add(parseValue(stream));
-            ch = stream.nextNonWhitespace();
-            if (ch == ']') {
-                return list;
-            } else if (ch != ',') {
-                throw new IllegalStateException("Invalid JSON array separator");
-            }
-        }
-    }
-
-    private static String parseString(final LenientStream stream) throws IOException {
-        final StringBuilder sb = new StringBuilder();
-        boolean escaped = false;
-        while (true) {
-            final int ch = stream.read();
-            if (ch == -1) {
-                throw new IllegalStateException("Unterminated string");
-            }
-            if (escaped) {
-                sb.append(decodeEscape(ch, stream));
-                escaped = false;
-            } else if (ch == '\\') {
-                escaped = true;
-            } else if (ch == '"') {
-                break;
-            } else {
-                sb.append((char) ch);
-            }
-        }
-        return sb.toString();
-    }
-
-    private static char decodeEscape(final int esc, final LenientStream stream) throws IOException {
-        return switch (esc) {
-            case '"' -> '"';
-            case '\\' -> '\\';
-            case '/' -> '/';
-            case 'b' -> '\b';
-            case 'f' -> '\f';
-            case 'n' -> '\n';
-            case 'r' -> '\r';
-            case 't' -> '\t';
-            case 'u' -> {
-                final char[] hex = new char[4];
-                for (int i = 0; i < 4; i++) {
-                    final int h = stream.read();
-                    if (h == -1) {
-                        throw new IllegalStateException("Incomplete unicode escape");
-                    }
-                    hex[i] = (char) h;
-                }
-                yield (char) Integer.parseInt(new String(hex), 16);
-            }
-            default -> (char) esc;
-        };
-    }
-
-    private static Object parsePrimitive(final LenientStream stream, final char first) throws IOException {
-        final StringBuilder token = new StringBuilder();
-        token.append(first);
-        while (true) {
-            final int ch = stream.read();
-            if (ch == -1 || ch == ',' || ch == '}' || ch == ']' || Character.isWhitespace(ch)) {
-                if (ch != -1) {
-                    stream.unread(ch);
-                }
-                break;
-            }
-            token.append((char) ch);
-        }
-        final String value = token.toString();
-        return switch (value) {
-            case "true" -> true;
-            case "false" -> false;
-            case "null" -> null;
-            default -> convertToPrimitive(value);
-        };
-    }
-
-    private static Object convertToPrimitive(final String value) {
-        if ("null".equals(value)) {
-            return null;
-        } else if ("true".equals(value) || "false".equals(value)) {
-            return Boolean.parseBoolean(value);
-        } else {
-            final Object result;
-            if (value.contains(".") || value.contains("e") || value.contains("E")) {
-                result = convertObj(value, Double.class);
-            } else {
-                result = convertObj(value, Long.class);
-            }
-            return result == null ? unescapeJson(value) : result;
-        }
-    }
-
-    private static class LenientStream {
-        private static final int PUSHBACK_SIZE = 16;
-        private final PushbackReader reader;
-        private final StringBuilder raw;
-        private final int rawLimit;
-
-        LenientStream(final Reader reader, final int rawLimit) {
-            this.reader = new PushbackReader(reader, PUSHBACK_SIZE);
-            this.rawLimit = rawLimit;
-            this.raw = new StringBuilder(Math.min(rawLimit, 1024));
-        }
-
-        int read() throws IOException {
-            final int ch = reader.read();
-            appendRaw(ch);
-            return ch;
-        }
-
-        void unread(final int ch) throws IOException {
-            if (ch != -1) {
-                reader.unread(ch);
-            }
-        }
-
-        int nextNonWhitespace() throws IOException {
-            int ch;
-            do {
-                ch = read();
-            } while (ch != -1 && Character.isWhitespace(ch));
-            return ch;
-        }
-
-        String rawText() {
-            return raw.toString();
-        }
-
-        private void appendRaw(final int ch) {
-            if (ch != -1 && raw.length() < rawLimit) {
-                raw.append((char) ch);
-            }
-        }
     }
 
     private JsonDecoder() {
